Switch to BEM and use ids only for querySelector
I'm enjoying the flow of BEM CSS, for right now. Had the iconic thought of "I just change this here, and it works how I wanted it to". Kind of taking advantage of the Cascade.

Played with fonts and settled on two old favorites. There's probably a better way to load them, but I don't know it yet.